+++ tests/testthat/test-edge-nn.R
@@ -1,4 +1,4 @@
-context("test-edge-dist")
+context("test-edge-nn")
 
 library(spatsoc)
 
@@ -36,7 +36,8 @@ test_that('column names must exist in DT', {
     edge_nn(
       DT,
       id = 'potato',
-      coords = c('X', 'Y')
+      coords = c('X', 'Y'),
+      timegroup = 'timegroup'
     ),
     'not present in input DT',
     fixed = FALSE
@@ -47,7 +48,8 @@ test_that('column names must exist in DT', {
     edge_nn(
       DT,
       id = 'ID',
-      coords = c('potatoX', 'potatoY')
+      coords = c('potatoX', 'potatoY'),
+      timegroup = 'timegroup'
     ),
     'not present in input DT',
     fixed = FALSE
@@ -59,7 +61,8 @@ test_that('column names must exist in DT', {
       DT,
       id = 'ID',
       coords = c('X', 'Y'),
-      splitBy = 'potato'
+      splitBy = 'potato',
+      timegroup = 'timegroup'
     ),
     'not present in input DT',
     fixed = FALSE
@@ -84,7 +87,8 @@ test_that('threshold correctly provided or error detected', {
   expect_silent(edge_nn(
     copyDT,
     id = 'ID',
-    coords = c('X', 'Y')
+    coords = c('X', 'Y'),
+    timegroup = NULL
   ))
 
   expect_error(edge_nn(DT, threshold = -10, id = 'ID'),
@@ -112,7 +116,8 @@ test_that('coords are correctly provided or error detected', {
     edge_nn(
       DT,
       id = 'ID',
-      coords = c('X', 'ID')
+      coords = c('X', 'ID'),
+      timegroup = NULL
     ),
     'coords must be numeric'
   )
@@ -205,4 +210,92 @@ test_that('returned IDs make sense', {
   expect_true(eDT[ID == NN, .N] == 0)
 
 
+})
+
+
+
+
+
+
+test_that('returned columns match', {
+  copyDT <- copy(DT)[, datetime := as.POSIXct(datetime)]
+  group_times(copyDT, datetime = 'datetime', threshold = '10 minutes')
+  eDT <- edge_nn(
+    copyDT,
+    id = 'ID',
+    coords = c('X', 'Y'),
+    timegroup = 'timegroup'
+  )
+
+  expect_true(all(c('ID', 'NN') %in% colnames(eDT)))
+  expect_false('distance' %in% colnames(eDT))
+
+
+  eDT <- edge_nn(
+    copyDT,
+    id = 'ID',
+    coords = c('X', 'Y'),
+    timegroup = 'timegroup',
+    returnDist = TRUE
+  )
+
+  expect_true(all(c('ID', 'NN', 'distance') %in% colnames(eDT)))
+
+
+  eDT <- edge_nn(
+    copyDT,
+    id = 'ID',
+    coords = c('X', 'Y'),
+    timegroup = 'timegroup',
+    returnDist = TRUE,
+    threshold = 1000
+  )
+
+  expect_true(all(c('ID', 'NN', 'distance') %in% colnames(eDT)))
+
+  eDT <- edge_nn(
+    copyDT,
+    id = 'ID',
+    coords = c('X', 'Y'),
+    timegroup = 'timegroup',
+    returnDist = FALSE,
+    threshold = 1000
+  )
+
+  expect_true(all(c('ID', 'NN') %in% colnames(eDT)))
+  expect_false('distance' %in% colnames(eDT))
+
+
+})
+
+
+test_that('distances returned are below threshold', {
+  copyDT <- copy(DT)[, datetime := as.POSIXct(datetime)]
+  group_times(copyDT, datetime = 'datetime', threshold = '10 minutes')
+  thresh <- 1000
+  eDT <- edge_nn(
+    copyDT,
+    id = 'ID',
+    coords = c('X', 'Y'),
+    timegroup = 'timegroup',
+    returnDist = TRUE,
+    threshold = thresh
+  )
+
+  expect_equal(eDT[distance > thresh, .N], 0)
+})
+
+test_that('NAs exist in NN when threshold provided', {
+  copyDT <- copy(DT)[, datetime := as.POSIXct(datetime)]
+  group_times(copyDT, datetime = 'datetime', threshold = '10 minutes')
+  thresh <- 1000
+  eDT <- edge_nn(
+    copyDT,
+    id = 'ID',
+    coords = c('X', 'Y'),
+    timegroup = 'timegroup',
+    threshold = thresh
+  )
+
+  expect_gt(eDT[is.na(NN), .N], 0)
 })